feat: implemented Water Jug Riddle feature
This commit includes deletion of the previously defined controller, service, and module files belonging to the old application. It replaces these files with new service, controller, and module implementations for the Water Jug Riddle problem. It also introduces a global cache module for the application.
Importantly, corresponding tests and data transfer objects for the service are added. In sum, a Water Jug Riddle operation with caching has been implemented.