feat: add Dockerfile and update package dependencies
This commit adds a new Dockerfile for the 'docs' application and updates the package dependencies in 'api' and 'web' applications. A new PDF file, 'Water bucket challenge 2024 BACKEND.pdf', has also been added to the repository. The newly added Dockerfile installs necessary dependencies, builds the 'docs' application and finally sets up a non-root user for running the application. The updated packages in both 'api' and 'web' applications aim to keep them up-to-date with the latest version.