feat: add server_url to docker-compose, use it in solve route
This commit adds a new environment variable, SERVER_URL, to the docker-compose file. This change allows the web app to connect to the API using the dynamic server URL dictated by docker-compose rather than a hard-coded one. In the `solve` route, a change was made to use this environment variable when making the API call instead of the previous hard-coded URL.