feat: add kubernetes deployment files and update build script
The commit introduces Kubernetes deployment files for the 'api' and 'web' applications. The build-and-run-docker-image.sh script has been updated to include the building of Docker images for 'api' and 'web' apps, and deploys them to Kubernetes. Istio installation and port forwarding for observability tools have also been added to the script.