feat: add horizontal auto scaling using a metrics-server
The main actions included here are adding the configuration for the metrics-server in the Kubernetes setup, improving the documentation in Swagger, and integrating an autoscaler for the Web deployment. The changes also include an update in the 'pnpm-lock.yaml' file. The metrics-server microservice was added to monitor other services in the Kubernetes cluster, which is an important improvement from a DevOps perspective. An example response was added in the Swagger documentation for solving the water jug riddle. Kubernetes deployment 'web.yaml' is now configured with 'HorizontalPodAutoscaler' for dynamic scaling of pods based on CPU utilization. Update the lockfile version and some dependencies in 'pnpm-lock.yaml'.